reorganize package structure for rest
