+++ src/main/java/org/jitsi/videobridge/rest/root/Application.java
@@ -14,18 +14,11 @@
  * limitations under the License.
  */
 
-package org.jitsi.videobridge.rest;
+package org.jitsi.videobridge.rest.root;
 
 import org.glassfish.jersey.server.*;
 import org.jitsi.utils.logging2.*;
-import org.jitsi.videobridge.rest.about.health.*;
-import org.jitsi.videobridge.rest.about.version.Version;
 import org.jitsi.videobridge.rest.binders.*;
-import org.jitsi.videobridge.rest.conferences.*;
-import org.jitsi.videobridge.rest.debug.*;
-import org.jitsi.videobridge.rest.mucclient.*;
-import org.jitsi.videobridge.rest.shutdown.Shutdown;
-import org.jitsi.videobridge.rest.stats.*;
 import org.osgi.framework.*;
 
 public class Application extends ResourceConfig
@@ -35,14 +28,15 @@ public class Application extends ResourceConfig
     public Application(BundleContext bundleContext)
     {
         register(new OsgiServiceBinder(bundleContext));
-        register(Stats.class);
-        register(Version.class);
-        register(Health.class);
-        register(Conferences.class);
+        packages("org.jitsi.videobridge.rest.root");
+//        register(Stats.class);
+//        register(Version.class);
+//        register(Health.class);
+//        register(Conferences.class);
         //TODO: add check for config prop
-        logger.info("Enabling shutdown via REST");
-        register(Shutdown.class);
-        register(Debug.class);
-        register(MucClient.class);
+//        logger.info("Enabling shutdown via REST");
+//        register(Shutdown.class);
+//        register(Debug.class);
+//        register(MucClient.class);
     }
 }
