+++ src/test/java/org/jitsi/videobridge/rest/VideobridgeRestResourceTest.java
@@ -18,10 +18,9 @@ package org.jitsi.videobridge.rest;
 
 import org.glassfish.jersey.test.*;
 import org.jitsi.videobridge.*;
+import org.jitsi.videobridge.rest.root.*;
 import org.jitsi.videobridge.util.*;
 import org.junit.*;
-
-import javax.ws.rs.core.*;
 
 import static org.mockito.Mockito.*;
