+++ src/test/java/org/jitsi/videobridge/rest/root/colibri/debug/DebugTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.jitsi.videobridge.rest.debug;
+package org.jitsi.videobridge.rest.root.colibri.debug;
 
 import org.eclipse.jetty.http.*;
 import org.glassfish.jersey.server.*;
